// src/domain/shared/enums/pokemon-type.enum.ts
export enum PokemonType {
  Normal = "normal",
  Fire = "fire",
  Water = "water",
  Grass = "grass",
  Electric = "electric",
  Ice = "ice",
  Fighting = "fighting",
  Poison = "poison",
  Ground = "ground",
  Flying = "flying",
  Psychic = "psychic",
  Bug = "bug",
  Rock = "rock",
  Ghost = "ghost",
  Dark = "dark",
  Dragon = "dragon",
  Steel = "steel",
  Fairy = "fairy",
}